+++ src/main/java/com/award/sy/entity/WalletLog.java
@@ -27,9 +27,9 @@ public class WalletLog extends Po{
 	
 	private Long user_id;//用户ID
 	
-	private BigDecimal change_money;//变动金额
+	private Double change_money;//变动金额
 	
-	private BigDecimal money;//变动后的金额
+	private Double money;//变动后的金额
 	
 	private String remark;//备注
 	
@@ -67,19 +67,19 @@ public class WalletLog extends Po{
 		this.user_id = user_id;
 	}
 
-	public BigDecimal getChange_money() {
+	public Double getChange_money() {
 		return change_money;
 	}
 
-	public void setChange_money(BigDecimal change_money) {
+	public void setChange_money(Double change_money) {
 		this.change_money = change_money;
 	}
 
-	public BigDecimal getMoney() {
+	public Double getMoney() {
 		return money;
 	}
 
-	public void setMoney(BigDecimal money) {
+	public void setMoney(Double money) {
 		this.money = money;
 	}
